dashboard_service/requirements.txt
# Dashboard Service Requirements
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
pydeck==0.8.0
requests==2.28.2
python-dotenv==1.0.0
numpy==1.24.3
